fix(eval): split test_harness.py's anchor tests into test_golden_anchors.py
PR #137's self-scan gate went red (diagnosis.hospital: architecture gate
failure) because the 8 new anchor tests pushed eval/test_harness.py to 486
loc / 44 defs, tripping the sentrux god-file ratchet
(loc > 800 || (functions > 25 && loc > 400); crates/code-intel-cli/src/
sentrux_gate.rs:794) -- the exact rule eval/arms.py's own module docstring
already documents as the reason it was split out of harness.py in the
first place.

Move GoldenAnchorTests and the two real-questions anchor checks into a new
eval/test_golden_anchors.py (189 loc / 11 defs); eval/test_harness.py is
now 356 loc / 35 defs. Both are comfortably under the ratchet regardless
of how "functions" is counted, since loc is the binding term for both
files (356 and 189, both well under the 400 floor the function-count
clause needs to even apply).

Verified two ways:
  - `code-intel.exe run execute --repo . --out ... --manifest
    orchestration/integrations.json ...` (the exact command
    .github/workflows/ci.yml's "Authoritative self-scan" step runs)
    locally: exit 0, diagnosis.hospital verdict "pass".
  - `python eval/test_harness.py` (26 passed) + `python eval/
    test_golden_anchors.py` (9 passed) = 35, same coverage as the 34 tests
    before the split plus one additional dedicated schema-shape test.

No CI job invokes eval/test_harness.py by explicit path anywhere (checked
.github/workflows/*.yml and tests/test_repository_layout.py) -- the red
gate was the self-scan's static LOC/function-count check on the file
itself, not a missed test invocation, so there was nothing to wire up
there.
