+++ eval/test_harness.py
@@ -41,7 +41,6 @@ def _load_harness():
 
 
 harness = _load_harness()
-import golden_anchors  # noqa: E402  (eval/ is on sys.path once harness.py has executed)
 
 
 def _fixture_questions():
@@ -257,113 +256,6 @@ class RangeHelperTests(unittest.TestCase):
         self.assertFalse(harness.span_fully_covered(1, 3, [(5, 12)]))
 
 
-class GoldenAnchorTests(unittest.TestCase):
-    """eval/golden_anchors.py: golden spans resolve by content, not by a
-    pinned absolute line number, so an unrelated refactor that moves code
-    cannot silently rot the benchmark (issue #93 golden-symbol-anchors;
-    proof case: PR #134 shifted crates/code-intel-cli/src/sentrux_gate.rs's
-    evaluate_rules from line 481 to line 537 without changing what it does,
-    which alone dropped the frozen "how" baseline from 5/5 to 3/5)."""
-
-    def _write_repo(self, tmp: Path, content: str) -> None:
-        (tmp / "target.py").write_text(content, encoding="utf-8")
-
-    def test_symbol_anchor_resolves_to_the_current_declaration_line(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            self._write_repo(root, "x = 1\n\n\ndef add(a, b):\n    return a + b\n")
-            span = golden_anchors.resolve_anchor(
-                root, "target.py", {"kind": "symbol", "symbol_kind": "function", "name": "add", "after": 1}
-            )
-            self.assertEqual(span, (4, 5))
-
-    def test_symbol_anchor_survives_a_simulated_line_shift(self):
-        # The whole point of this mechanism: inject blank lines above the
-        # anchored declaration (simulating an unrelated upstream edit, the
-        # exact shape of what PR #134 did to sentrux_gate.rs) and prove the
-        # resolved span follows the declaration to its new line rather than
-        # staying pinned to the old one.
-        before_shift = "x = 1\n\n\ndef add(a, b):\n    return a + b\n"
-        anchor = {"kind": "symbol", "symbol_kind": "function", "name": "add", "after": 1}
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            self._write_repo(root, before_shift)
-            original_span = golden_anchors.resolve_anchor(root, "target.py", anchor)
-            self.assertEqual(original_span, (4, 5))
-
-            shift = 56  # matches the magnitude of the real evaluate_rules regression
-            shifted = ("\n" * shift) + before_shift
-            self._write_repo(root, shifted)
-            shifted_span = golden_anchors.resolve_anchor(root, "target.py", anchor)
-
-        self.assertEqual(shifted_span, (4 + shift, 5 + shift))
-        self.assertNotEqual(shifted_span, original_span)
-
-    def test_snippet_anchor_resolves_by_literal_text_not_position(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            self._write_repo(
-                root,
-                "# filler\n# filler\n# a distinctive rationale phrase here\n# filler\n",
-            )
-            span = golden_anchors.resolve_anchor(
-                root,
-                "target.py",
-                {"kind": "snippet", "text": "distinctive rationale phrase", "before": 1, "after": 1},
-            )
-            self.assertEqual(span, (2, 4))
-
-    def test_missing_symbol_fails_loudly_instead_of_returning_a_wrong_span(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            self._write_repo(root, "def other(): pass\n")
-            with self.assertRaises(golden_anchors.AnchorError):
-                golden_anchors.resolve_anchor(
-                    root, "target.py", {"kind": "symbol", "symbol_kind": "function", "name": "add"}
-                )
-
-    def test_ambiguous_snippet_fails_loudly_instead_of_picking_one(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            self._write_repo(root, "marker\nmarker\n")
-            with self.assertRaises(golden_anchors.AnchorError):
-                golden_anchors.resolve_anchor(
-                    root, "target.py", {"kind": "snippet", "text": "marker"}
-                )
-
-    def test_resolve_question_spans_reports_the_question_as_broken_on_failure(self):
-        question = {
-            "id": "qx",
-            "golden": {
-                "spans": [
-                    {
-                        "path": "target.py",
-                        "anchor": {"kind": "symbol", "symbol_kind": "function", "name": "missing"},
-                    }
-                ]
-            },
-        }
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            self._write_repo(root, "def present(): pass\n")
-            resolved, error = golden_anchors.resolve_question_spans(root, question)
-        self.assertIsNone(resolved)
-        self.assertIn("qx", error)
-        self.assertIn("missing", error)
-
-    def test_resolve_question_spans_passes_through_legacy_absolute_spans_unchanged(self):
-        # eval/fixtures/questions.json's tiny synthetic questions still use
-        # the pre-migration absolute shape; resolution must be a no-op for
-        # those rather than requiring every consumer to migrate at once.
-        question = {
-            "id": "qy",
-            "golden": {"spans": [{"path": "target.py", "start_line": 3, "end_line": 5}]},
-        }
-        resolved, error = golden_anchors.resolve_question_spans(Path("/nonexistent"), question)
-        self.assertIsNone(error)
-        self.assertEqual(resolved, [{"path": "target.py", "start_line": 3, "end_line": 5}])
-
-
 class CorpusExclusionTests(unittest.TestCase):
     """Corpus exclusion (issue #93 self-reference pollution fix, #104): neither
     arm may ever read eval/, docs/archive/, or .out-of-scope/ as an answer
@@ -426,6 +318,11 @@ class CorpusExclusionTests(unittest.TestCase):
 
 
 class QuestionSchemaTests(unittest.TestCase):
+    """Basic golden.* schema shape. Anchor-specific shape and live-tree
+    resolution of the real question set are eval/test_golden_anchors.py's
+    job (RealQuestionsAnchorTests) -- split out for the same god-file-ratchet
+    reason eval/golden_anchors.py's own resolver tests live there."""
+
     def test_real_questions_file_has_twelve_valid_entries(self):
         questions = harness.load_questions(EVAL_DIR / "questions.json")
         self.assertEqual(len(questions), 12)
@@ -435,32 +332,6 @@ class QuestionSchemaTests(unittest.TestCase):
             self.assertIn(q["category"], {"how", "where", "why"})
             self.assertTrue(q["golden"]["keywords"])
             self.assertTrue(q["golden"]["spans"])
-            for span in q["golden"]["spans"]:
-                # eval/questions.json is fully migrated to content anchors
-                # (issue #93 golden-symbol-anchors): every real span must
-                # carry an 'anchor', not an absolute start_line/end_line.
-                self.assertIn("anchor", span, f"{q['id']}: span must be anchor-based: {span}")
-                self.assertIn(span["anchor"]["kind"], {"symbol", "snippet"})
-
-    def test_real_questions_anchors_all_resolve_against_the_live_tree(self):
-        # The whole point: every golden anchor in the real question set must
-        # resolve against the actual current repo, or it's a broken question
-        # (see GoldenAnchorTests / eval/golden_anchors.py). q01 and q02 are
-        # the one known, expected exception: PR #134 moved
-        # is_primary_invocation and the RAW_ROUTES table out of main.rs
-        # entirely, which is a real code move this resolver correctly
-        # refuses to paper over -- it must report exactly those two broken,
-        # and nothing else.
-        questions = {q["id"]: q for q in harness.load_questions(EVAL_DIR / "questions.json")}
-        repo_root = EVAL_DIR.parent
-        broken = {}
-        for qid, question in questions.items():
-            _, error = golden_anchors.resolve_question_spans(repo_root, question)
-            if error:
-                broken[qid] = error
-        self.assertEqual(
-            set(broken), {"q01", "q02"}, f"unexpected broken/resolved anchors: {broken}"
-        )
 
     def test_duplicate_question_id_is_rejected(self):
         fixture_questions = json.loads((FIXTURES_DIR / "questions.json").read_text(encoding="utf-8"))
